refacto(watermark): delete offloading watermark and change *json.RawMessage to json.RawMessage
