+++ repositories/dbmodels/db_watermark.go
@@ -10,14 +10,14 @@ import (
 )
 
 type DbWatermark struct {
-	Id            uuid.UUID        `db:"id"`
-	OrgId         *string          `db:"org_id"`
-	Type          string           `db:"type"`
-	WatermarkTime time.Time        `db:"watermark_time"`
-	WatermarkId   *string          `db:"watermark_id"`
-	CreatedAt     time.Time        `db:"created_at"` //nolint:tagliatelle
-	UpdatedAt     time.Time        `db:"updated_at"`
-	Params        *json.RawMessage `db:"params"`
+	Id            uuid.UUID       `db:"id"`
+	OrgId         *string         `db:"org_id"`
+	Type          string          `db:"type"`
+	WatermarkTime time.Time       `db:"watermark_time"`
+	WatermarkId   *string         `db:"watermark_id"`
+	CreatedAt     time.Time       `db:"created_at"` //nolint:tagliatelle
+	UpdatedAt     time.Time       `db:"updated_at"`
+	Params        json.RawMessage `db:"params"`
 }
 
 const TABLE_WATERMARKS = "watermarks"
